Abstract out the comm object into a plugin
This change add a MachineComPlugin type, and add the default usage of
that Plugin, the basic_comm.

This change adds in some of the needed platform logic for dealing with
comm plugin selection at printer connection time, but leaves the ui work
to be done.

AE-663
